avance con bar busqueda
