+++ src/app/dashboard/candidato/buscar-empleos/page.jsx
@@ -1,6 +1,7 @@
 import SearchFilters from "../buscar-empleos/search-filters";
 import Select from "../../../components/ui/Select";
 import JobCard from "../../../components/dashboard/JobCard";
+//import JobDetail from "@/app/components/dashboard/JobDetail";
 
 export default function BuscarEmpleoPage() {
     return(
@@ -22,11 +23,10 @@ export default function BuscarEmpleoPage() {
               </div>
             </div>
             <div>
-
                 <JobCard />
             </div>
             <div>
-                {/* Job Detail */}
+                {/* JobDetail */}
             </div>
         </div>
     );
