+++ src/app/dashboard/candidato/mis-empleos/page.jsx
@@ -1,11 +1,12 @@
 // Mis empleos Page
-//import JobCard from '@/app/components/dashboard/JobCard';
+
+import JobCard from "../../../components/dashboard/JobCard";
 
 export default function MisEmpleosPage() {
     return(
         <div>
             <h1>Mis Empleos</h1>
-            {/*<JobCard />*/}
+            <JobCard />
         </div>
     )
-} 
+}
